services/core/requirements.txt
aio-pika==9.4.3
aiormq==6.8.1
asgiref==3.8.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
idna==3.8
multidict==6.0.5
pamqp==3.3.0
pika==1.3.2
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.1
yarl==1.10.0
